Update to download and notebook

+++ download_pubchem_ftp.py
@@ -7,20 +7,26 @@ def main() :
     ftp.login("anonymous", "anonymous")
     ftp.cwd("/pubchem/Compound/CURRENT-Full/SDF")
     try :
-        os.mkdirs("~/media/data/pubchem/SDF")
+        os.mkdirs("/media/data/pubchem/SDF")
     except :
         print("Directory SDF already exists")
 
     filelist = ftp.nlst()
+    existingfiles = os.listdir("/media/data/pubchem/SDF")
     print("Downloading: {0} files".format(len(filelist)))
     i = 0
     for filename in filelist:
 
-        local_filename = os.path.join('SDF', filename)
+        local_filename = os.path.join('/media/data/pubchem/SDF', filename)
+        if filename in existingfiles:
+            i = i + 1
+            continue
+
         with open(local_filename, 'wb') as file :
             ftp.retrbinary('RETR ' + filename, file.write)
-        i = i + 1
-        print("Downloaded file {0} of {1}".format(i,len(filelist)))
+            i = i + 1
+
+        print("Processed file {0} of {1}".format(i,len(filelist)))
 
     ftp.quit()
